update readline
Check if response string is empty before pushing to history.

Add bool for saving to local history.
Add bool for saving to file history.
Pressing <esc> saves input to local history, but not file history.
Pressing <ctrl-c> does not save the input to any history.
Input with a leading space character saves input to local history, but
not file history.